V5
I have mainly cleaned the code and created new functions to refactor actions.
i have removed useless get functions.
I have removed some useless structs.
i have changed current time to remove the forbidden function

+++ philo/main.c
@@ -64,7 +64,10 @@ int	main(int argc, char **argv)
 	// 	print_logs(simulation);
 	// }
 	// print_simu_status(simulation);
-	destroy_mutexes(simulation);
-	free_simulation(simulation);
+	if (simulation != NULL)
+	{
+		destroy_mutexes(simulation);
+		free_simulation(simulation);
+	}
 	return (0);
 }
